Part 2 - modified program to run concurrent balls // Using thread instead timer // Changed structure of ballData and Ball classes from inherited to instance of ballData within Ball // moved border check to Manager from Ball // restructured functions that provide balls movement to check collison with other balls in real time (with lock) // added dictionary to prevent gliching of balls

+++ Model/InterfaceModelBall.cs
@@ -14,7 +14,7 @@ public abstract class InterfaceModelBall
 
     public abstract int CoordY { get; set; }
 
-    public abstract int Radius { get;}
+    public abstract double Radius { get;}
 
     public abstract void UpdateBall(Object s, PropertyChangedEventArgs e);
 }
